Fixup refactoring
