+++ nbdime/merging/__init__.py
@@ -5,7 +5,8 @@
 
 from __future__ import unicode_literals
 
-from .decisions import decide_merge, apply_decisions
+from .generic import decide_merge
+from .decisions import apply_decisions
 from .notebooks import merge_notebooks
 
-__all__ = ["decide_merge", "merge_notebooks"]
+__all__ = ["decide_merge", "merge_notebooks", "apply_decisions"]
